+++ README.md
@@ -1,10 +1,10 @@
 # Ultimate Crypto Widget #
 **Contributors:** heisenejii  
-**Tags:** crypto, finance, web3, blockchain, cryptocurrency, coinmarketcap, cryptocurrency price ticker, Crypto Widgets, Crypto Price Updates  
+**Tags:** crypto, cryptocurrency, finance, widget, real-time price ticker  
 **Requires at least:** 6.5.5  
-**Tested up to:** 6.6.1  
+**Tested up to:** 6.6.2  
 **Requires PHP:** 8.0  
-**Stable tag:** 0.1.2  
+**Stable tag:** 0.1.3  
 **License:** GPL-2.0-or-later  
 **Donate link:** http://ultimatecryptowidget.com  
 **License URI:** http://www.gnu.org/licenses/gpl-2.0.txt  
@@ -124,6 +124,10 @@ Enjoy the enhanced capabilities of the Ultimate Crypto Plugin Pro version!
      Payment plan details can be found on the [plugin page](https://ultimatecryptowidget.com/).
 
 ## Changelog ##
+
+### 0.1.3 ###
+* Support Wordpress version 6.6.2
+* Added support for Binance Stream API to fetch real-time cryptocurrency data using Binance API.
 
 ### 0.0.2 ###
 * Support Wordpress version 6.6.1
